# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grocery Delivery API
 * OpenAPI spec version: 1.0.0
 */
import type { Category } from "./category";
import type { ProductUnit } from "./productUnit";

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  description?: string | null;
  descriptionAr?: string | null;
  price: number;
  unit: ProductUnit;
  image?: string | null;
  categoryId?: number | null;
  category?: Category | null;
  featured: boolean;
  inStock: boolean;
  createdAt: Date;
}