# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grocery Delivery API
 * OpenAPI spec version: 1.0.0
 */
import type { ProductInputUnit } from "./productInputUnit";

export interface ProductInput {
  name: string;
  nameAr: string;
  description?: string;
  descriptionAr?: string;
  price: number;
  unit: ProductInputUnit;
  image?: string;
  categoryId?: number;
  featured?: boolean;
  inStock?: boolean;
}